Fix C64 startup (can't state-save "size_t").
git-svn-id: svn://messdev.no-ip.org/mess@11405 9849079f-6229-0410-a07f-84406ce305ba

+++ src/mess/machine/64h156.h
@@ -151,7 +151,7 @@ private:
 	// track
 	int m_side;								// disk side
 	UINT8 m_track_buffer[G64_BUFFER_SIZE];	// track data buffer
-	size_t m_track_len;						// track length
+	int m_track_len;						// track length
 	offs_t m_buffer_pos;					// current byte position within track buffer
 	int m_bit_pos;							// current bit position within track buffer byte
 	int m_bit_count;						// current data byte bit counter
